Changed: Remove unsplitVertically() method from commands/utils.
